Adds a noop_mode to concat::setup
Create a noop_mode parameter which allows the module user to elect to relax the
noop metaparameter on certain resources in the module.  Specifically, this mode
controls the noop metaparameter of every resource except for the final target
of a concat.  In 'strict' mode, all resources respect the --noop command line
argument.  In 'friendly' mode, file resources under concatdir and the execs
which cause concatenation are forced to false regardless of the --noop command
line argument.