Fix validate_cmd file resource parameter
The validate_cmd parameter is only valid on newer versions of puppet.
This module is supporting all 3.x versions of puppet. This commit only
applies the validate_cmd parameter to the file resource if it is
defined and adds documentation noting that it will not work on older
versons of puppet.