Lookup is_pe fact with getvar
If the system is not running PE, the $::is_pe fact will be undefined,
causing errors when run under strict variables mode. This patch uses
the getvar() function to look up the fact so that it won't fail if not
defined.

Alternative to https://github.com/puppetlabs/puppetlabs-concat/pull/270

This should fix the broken CI for the postgresql module.