Update spec tests for rspec-puppet 2.0
Update tests for rspec-puppet 2.0

This patch changes `should` to `catalogue` for raise_error matchers
since `should` has the same problems as `subject`, where it is not
evaluated the way it used to be. We also unpin rspec-puppet.

Guard validate_cmd test for older puppet versions

Test group membership only on newer puppet

The defined() function doesn't seem to work properly with puppet 3.4 and rspec.
defined() works on its own, rspec works on its own, but together they
determine that $gid is not defined and cause errors here. Work around
it by ignoring this check for older puppet version.