fix defaulted force behavior
With Concat 2.0.x we've deprecated the -force parameter and the resulting bug
caused concat files that have no content/fragments to overwrite the target file with an
empty file. Concat should only create an empty file if no file exists, and if the file
exists, then it should not overwrite it with an empty file.