(MODULES-2104) Workaround for PUP-1963
PUP-1963 captures a bug in which generated resources do not receive
dependency edges. Because concat_file works by generating a file
resource, until PUP-1963 is resolved notify and subscribe relationships
against a concat or concat_file resource will not work properly.

This commit works around the issue by causing the concat_file resource
itself to emit the event.

The generated file resource captures the desired state of the file.
Ideally, the concat_file resource could remain an effective
"metaresource" and simply pass through events, but that's currently not
possible. Therefore when the concat_file resource is evaluated this
commit causes it to "look ahead" to see whether or not the generated
file resource is insync. If the generated file resource is not insync,
the concat_file resource itself will declare that it is not insync, and
so emit the correct event. Resources which notify or subscribe to the
concat_file resource will correctly receive the emitted event.

The downside to this workaround is an extra event log any time the
concat_file resource is updated.