Set default tag on concat type
The concat / concat_fragment relationship is defined by a tag, so
without the correct tag being set on the concat resource itself it will
not collect the related fragments to build the final file content.

This commit adds a default value to the tag parameter which will
correctly set it to what the concat_fragment resources will have to
enable correct collection.