Associate fragments by path, not tag
Previously tag was declared as a parameter for concat_file and also
concat_fragment (also the name of a metaparameter), and the tag value
was used to associate concat_fragment resources with their concat_file
collector.

ANAICT this was an unnecessarily brittle means of associating
collections. Basing association on tags effectively relied on the end
user not using the tag parameter, which is not a safe assumption given
that "tag" is also a useful metaparameter.

This commit makes path the namevar of the concat_file type and changes
the association to be between the concat_file-path and
concat_fragment-target parameters.