Add tests, validation for native type converstion
This commit adds a spec file and validation methods to the concat type,
as well as porting the tests from the old defined type.

This porting is necessary since the concat resource has transitioned
from a defined type to a native type and so all the tests need to be
converted.

Tests not yet converted are in pending blocks.