(MAINT) Refactoring OSs based on puppet agent docs
Prior to this commit the metadata.json file for this module listed OSs that are not supported by puppet agent as supported.
This commit aims to refactor the metadata.json file to only list OSs that are supported by puppet agent.

In this commit:
- Support for Solaris 10 was dropped.
- Support for Windows 7 and 8.1 was dropped.
- Support for Windows Server 2008 R2 was dropped.
- Support for AIX 5.3 and 6.1 was dropped.

The list of supported OSs can be found here:
https://puppet.com/docs/pe/2021.7/supported_operating_systems.html